+++ src/main/java/com/github/ontio/core/sidechaingovernance/InflationParam.java
@@ -1,6 +1,7 @@
 package com.github.ontio.core.sidechaingovernance;
 
-import sun.jvm.hotspot.debugger.Address;
+
+import com.github.ontio.common.Address;
 
 public class InflationParam {
     public String sideChainId;
